re-enable integration tests
