--- packages/node-xmpp-client/.travis.yml
@@ -1,22 +1,0 @@
-sudo: required
-
-language: node_js
-
-node_js:
-  - '0.10'
-  - '0.12'
-  - '4'
-  - '5'
-  - '6'
-
-before_install:
-  - sudo apt-get update
-  - sudo apt-get install prosody # luarocks
-  - sudo service prosody stop
-  # - sudo luarocks install luabitop
-  - sudo cp test/resources/prosody.cfg.lua /etc/prosody/prosody.cfg.lua
-  - sudo cp test/resources/mod_websocket.lua /usr/lib/prosody/modules/
-
-before_script: npm install -g grunt-cli
-
-script: npm run full-test
