+++ packages/node-xmpp-client/Gruntfile.js
@@ -5,19 +5,6 @@ var helper = require('./test/helper')
 module.exports = function (grunt) {
   grunt.initConfig({
     pkg: grunt.file.readJSON('package.json'),
-    mochacli: {
-      unit: {
-        options: { files: [ './test/unit/**/*.js' ] }
-      },
-      integration: {
-        options: { files: [ './test/integration/**/*.js' ] }
-      },
-      options: {
-        reporter: 'spec',
-        ui: 'tdd',
-        timeout: 10000
-      }
-    },
     mocha_phantomjs: {
       all: ['test/browser/**/*.html']
     },
@@ -37,14 +24,12 @@ module.exports = function (grunt) {
   })
 
   // Load the plugins
-  grunt.loadNpmTasks('grunt-mocha-cli')
   grunt.loadNpmTasks('grunt-mocha-phantomjs')
   grunt.loadNpmTasks('grunt-contrib-connect')
 
   // Configure tasks
   grunt.registerTask('default', ['test'])
-  grunt.registerTask('test', ['mochacli:unit'])
-  grunt.registerTask('integration-test', ['mochacli:integration', 'test'])
+  grunt.registerTask('integration-test', ['test'])
   grunt.registerTask('browser-test', ['connect', 'prosody-start', 'mocha_phantomjs', 'prosody-stop'])
   grunt.registerTask('full-test', ['test', 'integration-test', 'browser-test'])
   grunt.registerTask('dev', ['connect'])
