+++ packages/node-xmpp-client/package.json
@@ -32,7 +32,6 @@
     "grunt": "^0.4.5",
     "grunt-cli": "^1.2.0",
     "grunt-contrib-connect": "0.11.2",
-    "grunt-mocha-cli": "^2.0.0",
     "grunt-mocha-phantomjs": "^2.0.0",
     "node-xmpp-server": "^2.1.3"
   },
