+++ packages/node-xmpp-client/test/authentication/external.js
@@ -2,7 +2,7 @@
 
 'use strict'
 
-var External = require('../../../../lib/authentication/external')
+var External = require('../../lib/authentication/external')
 
 var mech = new External()
