+++ packages/node-xmpp-client/test/sasl.js
@@ -7,7 +7,7 @@ var C2SServer = nodexmppserver.C2SServer
 var Stanza = require('node-xmpp-core').Stanza
 var net = require('net')
 var util = require('util')
-var Client = require('../../index')
+var Client = require('../index')
 var Plain = nodexmppserver.auth.Plain
 var XOAuth2 = nodexmppserver.auth.XOAuth2
 var DigestMD5 = nodexmppserver.auth.DigestMD5
