+++ test/integration/client-tcp.js
@@ -2,13 +2,13 @@
 
 'use strict'
 
-var Client = require('../../index')
+var Client = require('../../packages/node-xmpp-client')
 var helper = require('../helper')
-var Element = require('node-xmpp-core').Element
+var Stanza = Client.Stanza
 
 require('should')
 
-describe.skip('BOSH connections', function () {
+describe('Socket connections', function () {
   var jid = Math.random().toString(36).substring(7) + '@localhost'
   var password = 'password'
   var client = null
@@ -19,18 +19,16 @@ describe.skip('BOSH connections', function () {
   })
 
   afterEach(function (done) {
-    if (client) client.end()
     helper.stopServer(done)
+    if (client) client.end()
   })
 
   it('Can register an account', function (done) {
     client = new Client({
       jid: jid,
       password: password,
-      register: true,
-      bosh: {
-        url: 'http://localhost:5280/http-bind/'
-      }
+      host: 'localhost',
+      register: true
     })
     client.on('online', function (data) {
       var bareJid = data.jid.local + '@' + data.jid.domain
@@ -45,9 +43,7 @@ describe.skip('BOSH connections', function () {
     client = new Client({
       jid: jid,
       password: 'not ' + password,
-      bosh: {
-        url: 'http://localhost:5280/http-bind/'
-      }
+      host: 'localhost'
     })
     client.on('online', function () {
       done('Should not have connected')
@@ -64,9 +60,7 @@ describe.skip('BOSH connections', function () {
     client = new Client({
       jid: jid + '/' + resource,
       password: password,
-      bosh: {
-        url: 'http://localhost:5280/http-bind/'
-      }
+      host: 'localhost'
     })
     client.on('online', function (data) {
       var bareJid = data.jid.local + '@' + data.jid.domain
@@ -82,9 +76,7 @@ describe.skip('BOSH connections', function () {
     client = new Client({
       jid: jid,
       password: password,
-      bosh: {
-        url: 'http://localhost:5280/http-bind/'
-      }
+      host: 'localhost'
     })
     client.on('online', function (data) {
       var bareJid = data.jid.local + '@' + data.jid.domain
@@ -100,9 +92,7 @@ describe.skip('BOSH connections', function () {
     client = new Client({
       jid: jid,
       password: 'not ' + password,
-      bosh: {
-        url: 'http://localhost:5280/http-bind/'
-      },
+      host: 'localhost',
       register: true
     })
     client.on('online', function () {
@@ -120,14 +110,12 @@ describe.skip('BOSH connections', function () {
     client = new Client({
       jid: jid,
       password: password,
-      bosh: {
-        url: 'http://localhost:5280/http-bind/'
-      }
+      host: 'localhost'
     })
 
-    var ping = new Element(
-      'iq', { id: '123', type: 'get' }
-    ).c('ping', { xmlns: 'urn:xmpp:ping' })
+    var ping = new Stanza('iq', {
+      id: '123', type: 'get'
+    }).c('ping', { xmlns: 'urn:xmpp:ping' })
 
     client.on('online', function () {
       client.send(ping)
@@ -138,40 +126,14 @@ describe.skip('BOSH connections', function () {
     })
   })
 
-  it('Can send and receive stanzas', function (done) {
-    client = new Client({
-      jid: jid,
-      password: password,
-      bosh: {
-        url: 'http://localhost:5280/http-bind/'
-      }
-    })
-
-    var ping = new Element(
-      'iq', { id: '123', type: 'get' }
-    ).c('ping', { xmlns: 'urn:xmpp:ping' })
-
-    var counter = 0
-    client.on('online', function () {
-      client.send(ping)
-      client.on('stanza', function () {
-        ++counter
-        if (counter > 6) return done()
-        client.send(ping)
-      })
-    })
-  })
-
   it('Sends error for bad stanza', function (done) {
     client = new Client({
       jid: jid,
       password: password,
-      bosh: {
-        url: 'http://localhost:5280/http-bind/'
-      }
+      host: 'localhost'
     })
 
-    var badPing = new Element(
+    var badPing = new Stanza(
       'wtf', { id: '123', type: 'get' }
     ).c('ping', { xmlns: 'urn:xmpp:ping' })
 
@@ -185,14 +147,12 @@ describe.skip('BOSH connections', function () {
     })
   })
 
-  it('Errors when server is stopped', function (done) {
+  it("Can't connect when server is stopped", function (done) {
     helper.stopServer(function () {
       client = new Client({
         jid: jid,
         password: password,
-        bosh: {
-          url: 'http://localhost:5280/http-bind/'
-        }
+        host: 'localhost'
       })
       client.on('error', function (error) {
         error.message.should.match(/connect ECONNREFUSED/)
@@ -207,33 +167,14 @@ describe.skip('BOSH connections', function () {
     })
   })
 
-  it('Errors when providing bad BOSH url', function (done) {
+  it('Disconects', function (done) {
     client = new Client({
       jid: jid,
       password: password,
-      bosh: {
-        url: 'http://localhost:5280/bosh-bind/'
-      }
-    })
-    client.on('error', function (error) {
-      error.message.should.equal('HTTP status 404')
-      done()
-    })
-    client.on('online', function () {
-      done('Should not have connected')
-    })
-  })
-
-  it('Disconnects', function (done) {
-    client = new Client({
-      jid: jid,
-      password: password,
-      bosh: {
-        url: 'http://localhost:5280/http-bind/'
-      }
+      host: 'localhost'
     })
 
-    var ping = new Element(
+    var ping = new Stanza(
       'iq', { id: '123', type: 'get' }
     ).c('ping', { xmlns: 'urn:xmpp:ping' })
 
@@ -247,40 +188,16 @@ describe.skip('BOSH connections', function () {
     })
   })
 
-  describe('Prebind', function () {
-    it('Returns RID and SID', function (done) {
-      new Client({ // eslint-disable-line
-        jid: jid,
-        password: password,
-        preferred: 'PLAIN',
-        wait: '60',
-        bosh: {
-          url: 'http://localhost:5280/http-bind',
-          prebind: function (error, data) {
-            if (error) return done(error)
-            data.rid.should.exist
-            data.rid.length.should.be.above(5)
-            data.sid.should.exist
-            data.sid.length.should.be.above(5)
-            done()
-          }
-        }
-      })
-    })
-  })
-
   describe('Authentication', function () {
     it('Can connect using PLAIN authentication', function (done) {
       client = new Client({
         jid: jid,
         password: password,
-        bosh: {
-          url: 'http://localhost:5280/http-bind/'
-        },
+        host: 'localhost',
         preferred: 'PLAIN'
       })
 
-      var ping = new Element(
+      var ping = new Stanza(
         'iq', { id: '123', type: 'get' }
       ).c('ping', { xmlns: 'urn:xmpp:ping' })
 
@@ -297,13 +214,11 @@ describe.skip('BOSH connections', function () {
       client = new Client({
         jid: jid,
         password: password,
-        bosh: {
-          url: 'http://localhost:5280/http-bind/'
-        },
+        host: 'localhost',
         preferred: 'DIGEST-MD5'
       })
 
-      var ping = new Element(
+      var ping = new Stanza(
         'iq', { id: '123', type: 'get' }
       ).c('ping', { xmlns: 'urn:xmpp:ping' })
 
@@ -321,13 +236,10 @@ describe.skip('BOSH connections', function () {
         jid: '@anon.localhost',
         password: password,
         host: 'localhost',
-        bosh: {
-          url: 'http://localhost:5280/http-bind/'
-        },
         preferred: 'ANONYMOUS'
       })
 
-      var ping = new Element(
+      var ping = new Stanza(
         'iq', { id: '123', type: 'get' }
       ).c('ping', { xmlns: 'urn:xmpp:ping' })
